Lots of work on the XLNet modeler
Got it to work in the scheme where:
- Inputs are chunked with the expected label fed on to the backend as a series of masks
- Model tries to predict masks and loss is combined failure to do so
- Outputs are predicted at the last chunk, but training goes on through all chunks.

These nets dont seem to like to predict very long sequences without some sort of
recurrence to build that predictive sequence up, so this didnt work so well.

Loss drops fast and doesn't improve from there. I think that was basically the model
learning to fill the entire output space full of pads.